feat: change decimal to double for the taxes property

+++ Doppler.BillingUser/Model/AccountingEntry.cs
@@ -26,7 +26,7 @@ namespace Doppler.BillingUser.Model
         public string PaymentEntryType { get; set; }
         public int IdCurrencyType { get; set; }
         public decimal? CurrencyRate { get; set; }
-        public decimal? Taxes { get; set; }
+        public double? Taxes { get; set; }
         public string ErrorMessage { get; set; }
     }
 }
